requirements.txt
Pillow==11.0.0
tqdm==4.67.0